Struct oak_sev_guest::guest::AttestationRequest

#[repr(C)]
pub struct AttestationRequest {
    pub report_data: [u8; 64],
    pub vmpl: u32,
    /* private fields */
}

Request for an attestation report.

See Table 20 in https://www.amd.com/system/files/TechDocs/56860.pdf.

Fields
report_data: [u8; 64]

The custom data to be included in the attestation report.

vmpl: u32

The VM Protection Level (VMPL) to be used in the attestation report.

Must be greater or equal to the current VMPL and at most 3.

Implementations

impl AttestationRequest

pub const fn new() -> Self