Struct oak_sev_guest::guest::KeyResponse

#[repr(C)]
pub struct KeyResponse {
    pub status: u32,
    pub derived_key: [u8; 32],
    /* private fields */
}

Response containing the derived key.

See Table 19 in https://www.amd.com/system/files/TechDocs/56860.pdf.

Fields
status: u32

The status of the operation.

Use KeyResponse::get_status to try to convert this to a KeyStatus enum.

derived_key: [u8; 32]

The derived key if status is KeyStatus::Success.

Implementations

impl KeyResponse

pub fn get_status(&self) -> Option<KeyStatus>

Gets the status field as a KeyStatus enum if possible.

pub fn validate(&self) -> Result<(), &'static str>

Checks that all reserved bytes are zero and that the status field, the report size and the report format are all valid.