Module oak_sev_guest::instructions

Rust instruction wrappers for managing page state and interacting with the hypervisor.

Structs

PermissionMask

Permission mask used by the RMP.

RmpPermission

Representation of the RMP permission used by the RMPADJUST instruction.

Enums

InstructionError

The potential errors when calling the PVALIDATE or RMPADJUST instructions.

PageSize

The size of a memory page.

Validation

Whether a page is in the validated state or not.

Vmsa

Whether the page can be used as a VM save area.

Functions

pvalidate

Marks a page as validated or unvalidated in the RMP.

rmpadjust

Adjusts the permissions of a page in the RMP.

vmgexit

Unconditionally exits from the guest to the hypervisor.