Module oak_sev_guest::interrupts

Utilities for building interrupt handlers that need to modify the register contents, such as a #VC handler that handles VMM communication exceptions generated by CPUID calls.

The “x86-interrupt” calling convention does not allow modification of any general purpose registers, so the #VC handler requires a different approach.

Example usage:

#![feature(naked_functions)]
#![feature(asm_sym)]

use oak_sev_guest::interrupts::{
    mutable_interrupt_handler_with_error_code, MutableInterruptStackFrame,
};

mutable_interrupt_handler_with_error_code!(
    unsafe fn vmm_communication_exception_handler(
        stack_frame: &mut MutableInterruptStackFrame,
        error_code: u64,
    ) {
        // Provide fake values for CPUID.
        match error_code {
            // CPUID was intercepted.
            0x72 => {
                stack_frame.rax = 0;
                stack_frame.rbx = 0;
                stack_frame.rcx = 0;
                stack_frame.rdx = 0;
                // Advance RIP by the length of the CPUID instruction.
                stack_frame.rip += 2u64;
            }
            _ => panic!("#VC exception {:x?} not handled", error_code),
        };
    }
);

Re-exports

pub use mutable_interrupt_handler_with_error_code;

Structs

MutableInterruptStackFrame

A reference to the interrupt stack frame will be passed as the first argument of the inner handler function.