Trait oak_sev_guest::io::PortReader

pub trait PortReader<T> {
    // Required method
    unsafe fn try_read(&mut self) -> Result<T, &'static str>;
}

Reader that can be used to read values from a port.

Required Methods

unsafe fn try_read(&mut self) -> Result<T, &'static str>

Tries to read from the port.

Safety

This function is unsafe because port access could have unsafe side-effects.