Trait oak_sev_guest::io::PortWriter

pub trait PortWriter<T> {
    // Required method
    unsafe fn try_write(&mut self, value: T) -> Result<(), &'static str>;
}

Writer that can be used to write values to a port.

Required Methods

unsafe fn try_write(&mut self, value: T) -> Result<(), &'static str>

Tries to write a value to the port.

Safety

This function is unsafe because port access could have unsafe side-effects.