Module oak_sev_guest::secrets

This module contains structs that can be used to interpret the contents of the secrets page that is provisioned into the VM guest memory during SEV-SNP startup.

Structs

GuestReservedArea

Representation of the Secrets Page Guest Reserved Area.

SecretsPage

Representation of the secrets page.

Enums

Imi

Whether an initial migration image is enabled.

Constants

SECRETS_PAGE_MAX_VERSION

The mmaximum version of the secrets pages that we expect to receive.

SECRETS_PAGE_MIN_VERSION

The minimum version of the secrets pages that we expect to receive.

SECRETS_PAGE_SIZE

The size of the secrets page.