Module oak_sev_guest::vmsa

This module provides an implementation of the State Save Area (VMSA) that is used to store encrypted CPU-related state on AMD SEV-ES and SEV-SNP.

Structs

SegmentRegister

Representation of a segment register in 64-bit mode.

Vmsa

The State Save Area used for SEV-ES and SEV-SNP.

VmsaPage

The page containing the Save State Area used for SEV-ES and SEV-SNP.

Constants

VMSA_PAGE_SIZE

VMSA_SIZE

Functions

calculate_rdx_from_fms

When the CPU is reset, the value of RDX is set to the same value that would be returned in EAX when calling CPUID for leaf 1.