Struct oak_sev_guest::vmsa::Vmsa

#[repr(C)]
pub struct Vmsa {

    pub es: SegmentRegister,
    pub cs: SegmentRegister,
    pub ss: SegmentRegister,
    pub ds: SegmentRegister,
    pub fs: SegmentRegister,
    pub gs: SegmentRegister,
    pub gdtr: SegmentRegister,
    pub ldtr: SegmentRegister,
    pub idtr: SegmentRegister,
    pub tr: SegmentRegister,
    pub pl0_ssp: u64,
    pub pl1_ssp: u64,
    pub pl2_ssp: u64,
    pub pl3_ssp: u64,
    pub u_cet: u64,
    pub vmpl: u8,
    pub cpl: u8,
    pub efer: u64,
    pub xss: u64,
    pub cr4: u64,
    pub cr3: u64,
    pub cr0: u64,
    pub dr7: u64,
    pub dr6: u64,
    pub rflags: u64,
    pub rip: u64,
    pub dr0: u64,
    pub dr1: u64,
    pub dr2: u64,
    pub dr3: u64,
    pub dr0_addr_mask: u64,
    pub dr1_addr_mask: u64,
    pub dr2_addr_mask: u64,
    pub dr3_addr_mask: u64,
    pub rsp: u64,
    pub s_cet: u64,
    pub ssp: u64,
    pub isst_addr: u64,
    pub rax: u64,
    pub star: u64,
    pub lstar: u64,
    pub cstar: u64,
    pub sfmask: u64,
    pub kernel_gs_base: u64,
    pub sysenter_cs: u64,
    pub sysenter_esp: u64,
    pub sysenter_eip: u64,
    pub cr2: u64,
    pub g_pat: u64,
    pub dbgctl: u64,
    pub br_from: u64,
    pub br_to: u64,
    pub last_excp_from: u64,
    pub last_excp_to: u64,
    pub pkru: u32,
    pub tsc_aux: u32,
    pub guest_tsc_scale: u64,
    pub guest_tsc_offset: u64,
    pub reg_prot_nonce: u64,
    pub rcx: u64,
    pub rdx: u64,
    pub rbx: u64,
    pub rbp: u64,
    pub rsi: u64,
    pub rdi: u64,
    pub r8: u64,
    pub r9: u64,
    pub r10: u64,
    pub r11: u64,
    pub r12: u64,
    pub r13: u64,
    pub r14: u64,
    pub r15: u64,
    pub guest_exit_info_1: u64,
    pub guest_exit_info_2: u64,
    pub guest_exit_int_info: u64,
    pub guest_nrip: u64,
    pub sev_features: u64,
    pub vintr_ctrl: u64,
    pub guest_exit_code: u64,
    pub virtual_tom: u64,
    pub tlb_id: u64,
    pub pcpu_id: u64,
    pub event_inj: u64,
    pub xcr0: u64,
    pub x87_dp: u64,
    pub mxcsr: u32,
    pub x87_ftw: u16,
    pub x87_fsw: u16,
    pub x87_fcw: u16,
    pub x87_fop: u16,
    pub x87_ds: u16,
    pub x87_cs: u16,
    pub x87_rip: u64,
    pub fpreg_x87: [u8; 80],
    pub fpreg_xmm: [u8; 256],
    pub fpreg_ymm: [u8; 256],
    pub lbr_stack_state: [u8; 256],
    pub lbr_select: u64,
    pub ibs_fetch_ctl: u64,
    pub ibs_fetch_lin_addr: u64,
    pub ibs_op_ctl: u64,
    pub ibs_op_rop: u64,
    pub ibs_op_data: u64,
    pub ibs_op_data2: u64,
    pub ibs_op_data3: u64,
    pub ibs_dc_lin_addr: u64,
    pub bp_ibs_tgt_rip: u64,
    pub ic_ibs_extd_ctl: u64,
    /* private fields */

}

The State Save Area used for SEV-ES and SEV-SNP.

See table B-4 in https://www.amd.com/system/files/TechDocs/24593.pdf

Fields
es: SegmentRegister

The extra segment.

cs: SegmentRegister

The code segment.

ss: SegmentRegister

The stack segment.

ds: SegmentRegister

The data segment.

fs: SegmentRegister

General purpose segment.

gs: SegmentRegister

General purpose segment.

gdtr: SegmentRegister

Pointer to the global descriptor table.

ldtr: SegmentRegister

Pointer to the local descriptor table.

idtr: SegmentRegister

Pointer to the interrupt descriptor table.

tr: SegmentRegister

Pointer to a valid task state secgement in the global descriptor table.

pl0_ssp: u64

Shadow stack pointer for ring 0.

pl1_ssp: u64

Shadow stack pointer for ring 1.

pl2_ssp: u64

Shadow stack pointer for ring 2.

pl3_ssp: u64

Shadow stack pointer for ring 3.

u_cet: u64

Shadow stack MSR for user mode settings.

vmpl: u8

The current VM protection level.

cpl: u8

The current protection level (ring).

efer: u64

The extended feature enable register.

xss: u64

The IA32_XSS machine-specific register.

cr4: u64

Control Register 4.

cr3: u64

Control Register 3.

cr0: u64

Control Register 0.

dr7: u64

Debug register 7.

dr6: u64

Debug register 6.

rflags: u64

The 64-bit flags register.

rip: u64

The instruction pointer.

dr0: u64

Debug register 0.

dr1: u64

Debug register 1.

dr2: u64

Debug register 2.

dr3: u64

Debug register 3.

dr0_addr_mask: u64

Debug register 0 address mask.

dr1_addr_mask: u64

Debug register 1 address mask.

dr2_addr_mask: u64

Debug register 2 address mask.

dr3_addr_mask: u64

Debug register 3 address mask.

rsp: u64

The stack pointer.

s_cet: u64

Shadow stack MSR for supervisor mode settings.

ssp: u64

The shadow stack pointer.

isst_addr: u64

The address of the interrupt shadow stack.

rax: u64

The RAX register.

star: u64

The target address, code segment and stack segment when making a syscall in legacy mode.

lstar: u64

The target instruction pointer when making a syscall in 64-bit mode.

cstar: u64

The target instruction pointer when making a syscall in compatibility mode.

sfmask: u64

The syscall flag mask.

kernel_gs_base: u64

Register used by the SWAPGS instruction to swap the base of the GS general purpose segment.

sysenter_cs: u64

The code segment when using SYSENTER or SYSEXIT in legacy mode.

sysenter_esp: u64

The stack pointer when using SYSENTER or SYSEXIT in legacy mode.

sysenter_eip: u64

The instruction pointer when using SYSENTER or SYSEXIT in legacy mode.

cr2: u64

The CR2 control register.

g_pat: u64

The page attribute table for the guest.

dbgctl: u64

The value of the guest’s DebugCTL MSR.

br_from: u64

The value of the guest’s LastBranchFromIP MSR.

br_to: u64

The value of the guest’s LastBranchToIP MSR.

last_excp_from: u64

The value of the guest’s LastIntFromIP MSR.

last_excp_to: u64

The value of the guest’s LastIntToIP MSR.

pkru: u32

The protect keys rights register.

tsc_aux: u32

Additional information read by the RDTSC instruction.

guest_tsc_scale: u64

The guest’s time stamp counter scaling factor.

guest_tsc_offset: u64

The guests time stamp counter offset.

reg_prot_nonce: u64

Nonce used when VMSA register protection is enabled.

rcx: u64

The RCX register.

rdx: u64

The RDX register.

rbx: u64

The RBX register.

rbp: u64

The RBP register.

rsi: u64

The RSI register.

rdi: u64

The RDI register.

r8: u64

The R8 register.

r9: u64

The R9 register.

r10: u64

The R10 register.

r11: u64

The R11 register.

r12: u64

The R12 register.

r13: u64

The R13 register.

r14: u64

The R14 register.

r15: u64

The R14 register.

guest_exit_info_1: u64

The info 1 value for automatic exits.

guest_exit_info_2: u64

The info 2 value for automatic exits.

guest_exit_int_info: u64

The interrupt info value for automatic exits.

guest_nrip: u64

The next instruction pointer for automatic exits.

sev_features: u64

The guest-controlled SEV features that are selected.

vintr_ctrl: u64

The guest-controlled interrupt injection control settings.

guest_exit_code: u64

The exit code for automatic exits.

virtual_tom: u64

The virtual top-of-memory setting for the guest.

tlb_id: u64

Used by the hardware to track TLB information for the guest.

pcpu_id: u64

Used to control flushing of the guest TLB. Writing 0 to this value will cause the TLB to be flushed on the next VMRUN.

event_inj: u64

Field used for injecting events into the guest.

xcr0: u64

The XCR0 extended control register.

x87_dp: u64

The X87 floating point data pointer.

mxcsr: u32

The Media eXtensions Control and Status Register.

x87_ftw: u16

The X87 floating point tag word.

x87_fsw: u16

The X87 floating point status word.

x87_fcw: u16

The X87 floating point control word.

x87_fop: u16

The X87 floating point opcode.

x87_ds: u16

The X87 floating point data segment.

x87_cs: u16

The X87 floating point code segment.

x87_rip: u64

The X87 instruction pointer.

fpreg_x87: [u8; 80]

The X87 register state.

fpreg_xmm: [u8; 256]

The XMM register state.

fpreg_ymm: [u8; 256]

The YMM register state.

lbr_stack_state: [u8; 256]

The last branch record stack state.

lbr_select: u64

The value of the guest’s LastBranchStackSelect MSR.

ibs_fetch_ctl: u64

The value of the guest’s IbsFetchCtl MSR.

ibs_fetch_lin_addr: u64

The value of the guest’s IbsFetchCtl MSR.

ibs_op_ctl: u64

The value of the guest’s IbsOfCtl MSR.

ibs_op_rop: u64

The value of the guest’s IbsOpRip MSR.

ibs_op_data: u64

The value of the guest’s IbsOpData1 MSR.

ibs_op_data2: u64

The value of the guest’s IbsOpData2 MSR.

ibs_op_data3: u64

The value of the guest’s IbsOpData3 MSR.

ibs_dc_lin_addr: u64

The value of the guest’s IbsDcLinAd MSR.

bp_ibs_tgt_rip: u64

The value of the guest’s IbsBrTarget MSR.

ic_ibs_extd_ctl: u64

The value of the guest’s IbsFetchExtdCtl MSR.

Implementations

impl Vmsa

pub fn new_vcpu_boot(rdx: u64) -> Self

Creates a new instance of the VMSA that represents the state of a newly reset vCPU at boot time.

The current implementation tries to match the state of a new vCPU on QEMU running with KVM.

The value of RDX depends on the family, model and stepping of the CPU and can be calculated using calculate_rdx_from_fms.

For reference see https://github.com/qemu/qemu/blob/master/target/i386/cpu.h, https://github.com/qemu/qemu/blob/master/target/i386/cpu.c and https://github.com/qemu/qemu/blob/master/target/i386/kvm.c.