Crate oak_sev_guest

Rust wrappers for instructions and structs for use by SNP-active guests to manage page state and interact with the hypervisor and the secure processor.

Modules

ap_jump_table

cpuid

This module contains structs that can be used to interpret the contents of the CPUID page that is provisioned into the VM guest memory during SEV-SNP startup.

crypto

This module provides an optional implementation for encrypting and decrypting guest requests using the RustCrypto aes-gcm crate.

ghcb

This module contains an implementation of the guest-hypervisor communications block (GHCB) page that can be used for communicating with the hypervisor.

guest

Utilities for creating and processing SNP Guest Request protocol messages.

instructions

Rust instruction wrappers for managing page state and interacting with the hypervisor.

interrupts

Utilities for building interrupt handlers that need to modify the register contents, such as a #VC handler that handles VMM communication exceptions generated by CPUID calls.

io

msr

Rust implementations of the AMD SEV-SNP GHCB MSR protocol.

secrets

This module contains structs that can be used to interpret the contents of the secrets page that is provisioned into the VM guest memory during SEV-SNP startup.

vmsa

This module provides an implementation of the State Save Area (VMSA) that is used to store encrypted CPU-related state on AMD SEV-ES and SEV-SNP.

Macros

mutable_interrupt_handler_with_error_code

Traits

Translator

Memory address translation functions.