Module oak_sev_guest::ghcb

This module contains an implementation of the guest-hypervisor communications block (GHCB) page that can be used for communicating with the hypervisor.

Structs

Ghcb

The guest-host communications block.

GhcbProtocol

Implementation of the GHCB protocol using the wrapped GHCB data structure.

ValidBitmap

Flags indicating which fields in a specific GHCB instance are valid.

Constants

GHCB_PAGE_SIZE

The size of the GHCB page.

GHCB_PROTOCOL_VERSION

The version of the GHCB protocol and page layout that we expect to use.