Struct oak_sev_guest::ghcb::Ghcb

#[repr(C, align(4096))]
pub struct Ghcb {

    pub cpl: u8,
    pub xss: u64,
    pub dr7: u64,
    pub rax: u64,
    pub rcx: u64,
    pub rdx: u64,
    pub rbx: u64,
    pub sw_exit_code: u64,
    pub sw_exit_info_1: u64,
    pub sw_exit_info_2: u64,
    pub sw_scratch: u64,
    pub xcr0: u64,
    pub valid_bitmap: ValidBitmap,
    pub x87_state_gpa: u64,
    pub shared_buffer: [u8; 2032],
    pub protocol_version: u16,
    pub ghcb_usage: u32,
    /* private fields */

}

The guest-host communications block.

See: Table 3 in https://www.amd.com/system/files/TechDocs/56421-guest-hypervisor-communication-block-standardization.pdf

Fields
cpl: u8

The current privilege level of the executing code.

xss: u64

The value of the IA32_XSS model-specific reqister.

dr7: u64

The value of the DR7 debug register.

rax: u64

The value of the RAX register.

rcx: u64

The value of the RCX register.

rdx: u64

The value of the RDX register.

rbx: u64

The value of the RBX register.

sw_exit_code: u64

Guest-controlled exit code.

sw_exit_info_1: u64

Guest-controlled exit information 1.

sw_exit_info_2: u64

Guest-controlled exit information 2.

sw_scratch: u64

Guest-controlled additional information.

xcr0: u64

Value of the XCR0 extended control register.

valid_bitmap: ValidBitmap

Bitmap indicating which quadwords of the save state area are valid in the range from offset 0x000 through to offset 0x3ef.

A flag must be set for each field that is set as part of handling a non-automatic exit event. The value must also be checked on the return from the hypervisor.

x87_state_gpa: u64

The guest-physical address of the page that contains the x87-related saved state.

shared_buffer: [u8; 2032]

Area that can be used as a shared buffer for communicating additional information.

protocol_version: u16

The version of the GHCB protocol and page layout in use.

ghcb_usage: u32

The usage of the GHCB page. A value of 0 indicates the usage is in line with the definition as specified in this struct. Any other value indicates a hypervisor-defined usage.

Implementations

impl Ghcb

pub const fn new() -> Self

pub fn reset(&mut self)

Zeroes the entire GHCB.