Module oak_sev_guest::guest

Utilities for creating and processing SNP Guest Request protocol messages.

Structs

AttestationRequest

Request for an attestation report.

AttestationResponse

Response containing the attestation report.

AuthenticatedHeader

The authenticated subsection of the header used for an encrypted guest request message.

EcdsaPublicKey

An ECDSA public key.

GuestFieldFlags

Flags indicating allowed policy options.

GuestMessage

An encrypted guest message.

GuestMessageHeader

The header for an encrypted guest request message.

KeyRequest

Request for a derived key.

KeyResponse

Response containing the derived key.

KeySelectIter

An iterator over the variants of KeySelect

RootKeySelectIter

An iterator over the variants of RootKeySelect

Enums

AeadAlgorithm

The AEAD algorithm used for encryption.

EccCurve

The elliptic curve used.

KeySelect

The selected key to use for key derivation.

KeyStatus

The status of the report response.

MessageType

The type of message represented by the payload.

ReportStatus

The status of the report response.

RootKeySelect

The selected root key to use for key derivation.

Constants

CURRENT_ATTESTATION_VERSION

The currently supported attestation report version number.

CURRENT_HEADER_VERSION

The currently supported header version number.

CURRENT_MESSAGE_VERSION

The currently supported message version number.

GUEST_MESSAGE_SIZE

The size of a guest message, including the header and maximum payload size.

MAX_PAYLOAD_SIZE

The maximum payload size.

Traits

Message