Module oak_sev_guest::guest

Utilities for creating and processing SNP Guest Request protocol messages.

Structs

AttestationReport

A signed attestation report.

AttestationReportData

The data contained in an attestation report.

AttestationRequest

Request for an attestation report.

AttestationResponse

Response containing the attestation report.

AuthenticatedHeader

The authenticated subsection of the header used for an encrypted guest request message.

EcdsaPublicKey

An ECDSA public key.

EcdsaSignature

An ECDSA signature.

GuestFieldFlags

Flags indicating allowed policy options.

GuestMessage

An encrypted guest message.

GuestMessageHeader

The header for an encrypted guest request message.

GuestPolicy

The required policy for a guest to run.

KeyRequest

Request for a derived key.

KeyResponse

Response containing the derived key.

KeySelectIter

An iterator over the variants of Self

PlatformInfo

Information on the platform configuration.

PolicyFlags

Flags indicating allowed policy options.

RootKeySelectIter

An iterator over the variants of Self

TcbVersion

The version of all the components in the Trusted Computing Base (TCB).

Enums

AeadAlgorithm

The AEAD algorithm used for encryption.

AuthorKey

Whether the author key digest is included in the report.

EccCurve

The elliptic curve used.

KeySelect

The selected key to use for key derivation.

KeyStatus

The status of the report response.

MessageType

The type of message represented by the payload.

ReportStatus

The status of the report response.

RootKeySelect

The selected root key to use for key derivation.

SigningAlgorithm

The signing algorithm used for the report signature.

Constants

CURRENT_ATTESTATION_VERSION

The currently supported attestation report version number.

CURRENT_HEADER_VERSION

The currently supported header version number.

CURRENT_MESSAGE_VERSION

The currently supported message version number.

GUEST_MESSAGE_SIZE

The size of a guest message, including the header and maximum payload size.

MAX_PAYLOAD_SIZE

The maximum payload size.

Traits

Message