Module oak_sev_guest::io

Structs

GhcbIoFactory

A factory for creating port readers and writers that use the GHCB IOIO protocol.

GhcbIoPort

GHCB-based wrapper for a single IO port.

RawIoPortFactory

Factory for creating port reader and writers that perform direct IO operations.

Enums

PortFactoryWrapper

Wrapper implementation that can either create IO ports that perform direct IO or IO ports that use the GHCB IOIO protocol.

PortWrapper

Traits

IoPortFactory

Factory for instantiating IO port readers and writers.

PortReader

Reader that can be used to read values from a port.

PortWriter

Writer that can be used to write values to a port.

Type Aliases

StaticGhcbIoPort

An IO port reader and writer implementation that uses the GHCB IOIO protocol, static references and a spinlock for synchronisation.