Trait oak_sev_guest::io::IoPortFactory

pub trait IoPortFactory<'a, T, R: PortReader<T> + 'a, W: PortWriter<T> + 'a> {
    // Required methods
    fn new_reader(&self, port: u16) -> R;
    fn new_writer(&self, port: u16) -> W;
}

Factory for instantiating IO port readers and writers.

The typical usage is to either create raw instances that peform direct IO on the ports, or instances that use the GHCB IOIO protocol.

Required Methods

fn new_reader(&self, port: u16) -> R

Creates a new IO port reader instance.

fn new_writer(&self, port: u16) -> W

Creates a new IO port writer instance.