Type Alias oak_sev_guest::io::StaticGhcbIoPort

pub type StaticGhcbIoPort = GhcbIoPort<'static, RawSpinlock, GhcbProtocol<'static, Ghcb>, Ghcb>;

An IO port reader and writer implementation that uses the GHCB IOIO protocol, static references and a spinlock for synchronisation.

Aliased Type

struct StaticGhcbIoPort { /* private fields */ }